requirements.txt
rdkit-pypi==2020.9.5.2
seaborn==0.11.2
hydra-core
grpcio
numba
sklearn
